requirements.txt
openai-whisper==20231117
moviepy==2.1.2
ffmpeg-python==0.2.0
python-dotenv==1.0.1